appmodel: extract AppObjectGenerator base class to its own module
