+++ modules/appmodel/CMakeLists.txt
@@ -9,6 +9,7 @@ set (APPMODEL_SOURCES
     appmodel-parser.c
     appmodel-plugin.c
     appmodel-context.c
+    app-object-generator.c
     app-parser-generator.c
     application.c
 )
