+++ modules/appmodel/app-parser-generator.c
@@ -22,108 +22,8 @@
  *
  */
 
-#include "app-parser-generator.h"
+#include "app-object-generator.h"
 #include "appmodel.h"
-
-#include <string.h>
-
-typedef struct _AppObjectGenerator AppObjectGenerator;
-
-struct _AppObjectGenerator
-{
-  CfgBlockGenerator super;
-  gboolean (*parse_arguments)(AppObjectGenerator *self, CfgArgs *args, const gchar *reference);
-  void (*generate_config)(AppObjectGenerator *self, GlobalConfig *cfg, GString *result);
-  const gchar *included_apps;
-  const gchar *excluded_apps;
-  gboolean is_parsing_enabled;
-};
-
-static gboolean
-_is_application_included(AppObjectGenerator *self, const gchar *app_name)
-{
-  /* include everything if we don't have the option */
-  if (!self->included_apps)
-    return TRUE;
-  return strstr(self->included_apps, app_name) != NULL;
-}
-
-static gboolean
-_is_application_excluded(AppObjectGenerator *self, const gchar *app_name)
-{
-  if (!self->excluded_apps)
-    return FALSE;
-  return strstr(self->excluded_apps, app_name) != NULL;
-}
-
-static gboolean
-_parse_auto_parse_arg(AppObjectGenerator *self, CfgArgs *args, const gchar *reference)
-{
-  const gchar *v = cfg_args_get(args, "auto-parse");
-
-  if (v)
-    self->is_parsing_enabled = cfg_process_yesno(v);
-  else
-    self->is_parsing_enabled = TRUE;
-  return TRUE;
-}
-
-static gboolean
-_parse_auto_parse_exclude_arg(AppObjectGenerator *self, CfgArgs *args, const gchar *reference)
-{
-  const gchar *v = cfg_args_get(args, "auto-parse-exclude");
-  if (!v)
-    return TRUE;
-  self->excluded_apps = g_strdup(v);
-  return TRUE;
-}
-
-static gboolean
-_parse_auto_parse_include_arg(AppObjectGenerator *self, CfgArgs *args, const gchar *reference)
-{
-  const gchar *v = cfg_args_get(args, "auto-parse-include");
-  if (!v)
-    return TRUE;
-  self->included_apps = g_strdup(v);
-  return TRUE;
-}
-
-
-static gboolean
-app_object_generator_parse_arguments_method(AppObjectGenerator *self, CfgArgs *args, const gchar *reference)
-{
-  g_assert(args != NULL);
-
-  if (!_parse_auto_parse_arg(self, args, reference))
-    return FALSE;
-  if (!_parse_auto_parse_exclude_arg(self, args, reference))
-    return FALSE;
-  if (!_parse_auto_parse_include_arg(self, args, reference))
-    return FALSE;
-  return TRUE;
-}
-
-static gboolean
-_generate(CfgBlockGenerator *s, GlobalConfig *cfg, gpointer args, GString *result, const gchar *reference)
-{
-  AppObjectGenerator *self = (AppObjectGenerator *) s;
-  CfgArgs *cfgargs = (CfgArgs *)args;
-
-  if (!self->parse_arguments(self, cfgargs, reference))
-    return FALSE;
-
-  self->generate_config(self, cfg, result);
-
-  return TRUE;
-}
-
-void
-app_object_generator_init_instance(AppObjectGenerator *self, gint context, const gchar *name)
-{
-  cfg_block_generator_init_instance(&self->super, context, name);
-  self->super.generate = _generate;
-  self->parse_arguments = app_object_generator_parse_arguments_method;
-}
 
 /* app-parser() */
 
@@ -220,10 +120,10 @@ _generate_application(Application *app, gpointer user_data)
   if (strcmp(self->topic, app->super.instance) != 0)
     return;
 
-  if (!_is_application_included(&self->super, app->super.name))
+  if (!app_object_generator_is_application_included(&self->super, app->super.name))
     return;
 
-  if (_is_application_excluded(&self->super, app->super.name))
+  if (app_object_generator_is_application_excluded(&self->super, app->super.name))
     return;
 
   if (self->first_app_generated)
